support AWS S3

+++ backup.sh
@@ -3,6 +3,24 @@
 set -e
 set -o pipefail
 
+if [ "${S3_BUCKET}" = "**None**" ]; then
+  METHOD=FS
+else
+  METHOD=S3
+fi
+echo METHOD=${METHOD}
+
+if [ "${METHOD}" = "S3" && "${S3_ACCESS_KEY_ID}" = "**None**" ]; then
+  echo "You need to set the S3_ACCESS_KEY_ID environment variable."
+  exit 1
+fi
+
+if [ "${METHOD}" = "S3" && "${S3_SECRET_ACCESS_KEY}" = "**None**" ]; then
+  echo "You need to set the S3_SECRET_ACCESS_KEY environment variable."
+  exit 1
+fi
+
+
 if [ "${POSTGRES_DATABASE}" = "**None**" ]; then
   echo "You need to set the POSTGRES_DATABASE environment variable."
   exit 1
@@ -28,6 +46,17 @@ if [ "${POSTGRES_PASSWORD}" = "**None**" ]; then
   exit 1
 fi
 
+if [ "${S3_ENDPOINT}" == "**None**" ]; then
+  AWS_ARGS=""
+else
+  AWS_ARGS="--endpoint-url ${S3_ENDPOINT}"
+fi
+
+# env vars needed for aws tools
+export AWS_ACCESS_KEY_ID=$S3_ACCESS_KEY_ID
+export AWS_SECRET_ACCESS_KEY=$S3_SECRET_ACCESS_KEY
+export AWS_DEFAULT_REGION=$S3_REGION
+
 export PGPASSWORD=$POSTGRES_PASSWORD
 POSTGRES_HOST_OPTS="-h $POSTGRES_HOST -p $POSTGRES_PORT -U $POSTGRES_USER $POSTGRES_EXTRA_OPTS"
 
@@ -41,10 +70,15 @@ else
   tar cvzf backup.tgz dump.sql ${DATA_FOLDERS_TO_BACKUP}
 fi
 
-echo "Copy achive to /backup"
-cp backup.tgz /backup/${POSTGRES_DATABASE}_$(date +"%Y-%m-%dT%H:%M:%SZ").tgz
-echo "Remove oldest backups"
-ls -tp /backup/*.tgz | grep -v '/$' | tail -n +${DAYS_HISTORY}
-ls -tp /backup/*.tgz | grep -v '/$' | tail -n +${DAYS_HISTORY} | xargs -I {} rm -- {}
+if [ "${METHOD}" = "S3" ]; then
+  echo "Uploading achive to $S3_BUCKET"
+  cat backup.tgz | aws $AWS_ARGS s3 cp - s3://$S3_BUCKET/$S3_PREFIX/${POSTGRES_DATABASE}_$(date +"%Y-%m-%dT%H:%M:%SZ").tgz || exit 2
+else
+  echo "Copy achive to /backup"
+  cp backup.tgz /backup/${POSTGRES_DATABASE}_$(date +"%Y-%m-%dT%H:%M:%SZ").tgz
+  echo "Remove oldest backups"
+  ls -tp /backup/*.tgz | grep -v '/$' | tail -n +${DAYS_HISTORY}
+  ls -tp /backup/*.tgz | grep -v '/$' | tail -n +${DAYS_HISTORY} | xargs -I {} rm -- {}
+fi
 
 echo "DB backuped successfully"
